Extract controller output type
