+++ apps/src/core/organisms/brains/DuckNeuralNetRecurrentBrainV2.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "core/organisms/DuckBrain.h"
+#include "core/organisms/brains/ControllerOutput.h"
 #include "core/organisms/brains/Genome.h"
 #include "core/organisms/evolution/GenomeLayout.h"
 
@@ -13,17 +14,6 @@ class Duck;
 
 class DuckNeuralNetRecurrentBrainV2 : public DuckBrain {
 public:
-    struct ControllerOutput {
-        float x = 0.0f;
-        float y = 0.0f;
-        bool a = false;
-        bool b = false;
-        float xRaw = 0.0f;
-        float yRaw = 0.0f;
-        float aRaw = 0.0f;
-        float bRaw = 0.0f;
-    };
-
     DuckNeuralNetRecurrentBrainV2();
     explicit DuckNeuralNetRecurrentBrainV2(const Genome& genome);
     explicit DuckNeuralNetRecurrentBrainV2(uint32_t seed);
